requirements.txt
pandas
pyttsx3
openai
whisper-openai
matplotlib
ffmpeg-python
PyAudio
SpeechRecognition
seaborn
plotly
BeautifulSoup4
lxml